This commit integrates 577 commits from mainstream TileLang (since July 21, 2025)
with 135 TileScale distributed computing commits.

## Key Features from Mainstream TileLang

- **TVM-FFI API Modernization**: Updated to new TVM FFI macros for ~100x faster
  Python-C++ interop (TVM_FFI_REGISTER_GLOBAL, PackedArgs, etc.)
- **TileOperator Refactoring**: Renamed Operator to TileOperator with TVM reflection
- **New GPU Support**: SM70, SM100 (Blackwell), SM120 intrinsics
- **CuTeDSL Backend**: Alternative NVIDIA CuTe-based code generation
- **Sparse GEMM**: Sparse tensor core operations
- **Language v2 Frontend**: New language frontend improvements
- **Z3 SMT Solver**: Enhanced expression proving for bounds checking

## TileScale Distributed Features Preserved

- **Distributed TileOperators**: PutOp, GetOp, StOp, LdOp, WaitOp, BarrierBlocksOp
- **Language Primitives**: warp_any, warp_all, ld, st, shuffle_elect
- **NVSHMEM Integration**: RDC compilation, nvshmem_host/device linking
- **CUDA Templates**: ldst.h, distributed.h, sync.h (restored)
- **Multi-GPU Examples**: AllGather, All-to-All, Pre/Post-Attention collectives

## Execution Backend Architecture

- **Automatic Selection**: Auto-detects NVSHMEM mode and selects cython backend
- **Backend Options**: tvm_ffi (default), cython (NVSHMEM), nvrtc, cutedsl
- **NVSHMEM Path Discovery**: Auto-discovers from env, 3rdparty, or pip package

## Test Results

- Language tests: 184 passed, 17 skipped
- Kernel tests: 27 passed, 6 skipped
- Flash attention: 9 passed, 5 skipped
- Distributed examples: 5/5 tested passing (allgather, simple_shift, all_to_all,
  pre_attn_all2all, post_attn_all2all_transpose)

## Documentation

See docs/merge_tilescale/ for detailed documentation:
- BUILD_AND_RUN.md: Build and test instructions
- MERGE_RATIONALE.md: Why this merge is needed
- MERGE_ANALYSIS.md: Feature status and analysis
- EXEC_BACKEND_ANALYSIS.md: Backend architecture details

+++ examples/quickstart.py
@@ -6,13 +6,12 @@ import tilelang.language as T
 # target currently can be "cuda" or "hip" or "cpu".
 # if not specified, it will be inferred from the input tensors during compile time
 @tilelang.jit
-def matmul(M, N, K, block_M, block_N, block_K, dtype="float16", accum_dtype="float"):
-
+def matmul(M, N, K, block_M, block_N, block_K, dtype=T.float16, accum_dtype=T.float32):
     @T.prim_func
     def matmul_relu_kernel(
-            A: T.Tensor((M, K), dtype),
-            B: T.Tensor((K, N), dtype),
-            C: T.Tensor((M, N), dtype),
+        A: T.Tensor((M, K), dtype),
+        B: T.Tensor((K, N), dtype),
+        C: T.Tensor((M, N), dtype),
     ):
         # Initialize Kernel Context
         with T.Kernel(T.ceildiv(N, block_N), T.ceildiv(M, block_M), threads=128) as (bx, by):
@@ -55,10 +54,9 @@ block_M = 128
 block_N = 128
 block_K = 32
 
-# 1. Define the kernel (matmul) and compile/lower it into an executable module
+# Define the kernel (matmul) and compile/lower it into an executable module
 matmul_relu_kernel = matmul(M, N, K, block_M, block_N, block_K)
-
-# 3. Test the kernel in Python with PyTorch data
+# Test the kernel in Python with PyTorch data
 import torch
 
 # Create random input tensors on the GPU
@@ -78,7 +76,7 @@ torch.testing.assert_close(c, ref_c, rtol=1e-2, atol=1e-2)
 print("Kernel output matches PyTorch reference.")
 
 # 4. Retrieve and inspect the generated CUDA source (optional)
-# cuda_source = jit_kernel.get_kernel_source()
+# cuda_source = matmul_relu_kernel.get_kernel_source()
 # print("Generated CUDA kernel:\n", cuda_source)
 
 # 5.Profile latency with kernel
