+++ testing/python/issue/test_tilelang_issue_814.py
@@ -5,12 +5,11 @@ import torch
 
 
 @tilelang.jit
-def _tmp_var_kernel(N, block_N, dtype="float"):
-
+def _tmp_var_kernel(N, block_N, dtype=T.float32):
     @T.prim_func
     def kernel(
-            A: T.Tensor((N,), dtype),
-            B: T.Tensor((N,), dtype),
+        A: T.Tensor((N,), dtype),
+        B: T.Tensor((N,), dtype),
     ):
         with T.Kernel(T.ceildiv(N, block_N), threads=128) as bx:
             for i in T.Parallel(block_N):
